OpenBrowserWindowApi backfill
Name of the requirement-set: OpenBrowserWindowApi
Platforms: Windows, iOS, Mac. 
Versions: For build versions for Windows/Mac/iOS: in the platform it’s been implemented for years. So, a older versions of each platform is picked as a way to allow this to work. We do not have a specific build# when this was implemented. If we run into any quality issues with that, we'll adjust accordingly